+++ scripts/unixes/build_remaken_project.sh
@@ -91,7 +91,6 @@ pushd ${BUILDROOTFOLDER}/${MODE}/debug
 `${QMAKE_PATH}/qmake ${PROJECTROOT}/${PROJECTNAME}.pro -spec ${QMAKE_SPEC} ${QMAKE_MODE} CONFIG+=debug CONFIG+=x86_64 CONFIG+=qml_debug CONFIG+=XPCF_DISABLE_ATTRIBUTES && /usr/bin/make qmake_all`
 RESULT=$?
 if [[ ! ${RESULT} -eq 127 ]]; then
-	echo "pouet"
  	exit ${RESULT}
 fi
 echo "===> qmake done"